Refactor Nina`s tests
