+++ tests/pages/primary/auth_page.py
@@ -1,0 +1,23 @@
+from tests.pages.primary.component import Component
+from tests.pages.primary.page import Page
+
+
+class AuthPage(Page):
+    @property
+    def form(self):
+        return AuthForm(self.driver)
+
+
+class AuthForm(Component):
+    LOGIN = '//input[@name="st.email"]'
+    PASSWORD = '//input[@name="st.password"]'
+    SUBMIT = '//input[@data-l="t,sign_in"]'
+
+    def set_login(self, login):
+        self.driver.find_element_by_xpath(self.LOGIN).send_keys(login)
+
+    def set_password(self, pwd):
+        self.driver.find_element_by_xpath(self.PASSWORD).send_keys(pwd)
+
+    def submit(self):
+        self.driver.find_element_by_xpath(self.SUBMIT).click()
